consolidated all the child collection linq tests in one place
